requirements.txt
#txt2img model
diffusers==0.10.2
accelerate==0.12.0
safetensors==0.5.0
###
#composition model
academictorrents==2.3.3
albumentations==1.3.0
einops==0.3.0
imageio==2.9.0
more_itertools==10.2.0
natsort==8.4.0
numpy==1.21.6
omegaconf==2.3.0
#opencv_python==4.1.2.30
opencv_python_headless==4.7.0.72
packaging==23.2
Pillow==10.1.0
pudb==2019.2
pytorch_lightning==1.9.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.9.1
setuptools==65.6.3
streamlit==1.20.0
torch==1.10.1
torchmetrics==1.2.0
torchvision==0.11.2
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.10.0
scikit-image==0.19.3
taming-transformers-rom1504==0.0.5
###
#yolo
gitpython==3.1.30
matplotlib==3.3
#numpy>=1.23.5 #diff
opencv-python>=4.1.1
#pillow>=10.3.0 #diff
psutil  # system resources
#PyYAML>=5.3.1
#requests>=2.32.2 #diff
#scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
#torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
#torchvision>=0.9.0
#tqdm>=4.66.3 #diff
ultralytics>=8.2.34
pandas==1.1.4
###
#others
huggingface_hub==0.11.0
windows-curses